nuke widget control; add AutoSize wrapper component
