+++ packages/itwin/tree-widget/src/tree-widget-react.ts
@@ -1,10 +1,10 @@
 /*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
 
+export * from "./types";
 export * from "./TreeWidget";
-export * from "./components/TreeWidgetControl";
 export * from "./components/TreeWidgetComponent";
 export * from "./components/header/TreeHeader";
 export * from "./components/trees";
